Moves women in cs to utils. Updates logic so that only weekdays are included. Adds records to data. Updates brightness for background image.
