+++ client/src/App.css
@@ -17,6 +17,6 @@
   color: #61dafb;
 }
 
-.brightness-05 {
-  filter: brightness(.05);
+.brightness-08 {
+  filter: brightness(.08);
 }
