+++ client/src/checkIn/Confirmation.js
@@ -1,8 +1,7 @@
 import { useState, useEffect } from 'react'
 import axios from 'axios'
-import { womanInCS } from './womenInCS'
 import MoodChart from '../charts/MoodChart'
-import { server, dailyImage } from '../utils'
+import { server, dailyImage, womanInCS } from '../utils/utils'
 
 console.log(server)
 
@@ -41,7 +40,7 @@ const Confirmation = () => {
       <img
         src={require(`../images/${dailyImage}`)}
         alt=""
-        className="opacity-50 brightness-05 absolute inset-0 -z-10 h-full w-full object-cover object-right md:object-center" />
+        className="opacity-50 brightness-08 absolute inset-0 -z-10 h-full w-full object-cover object-right md:object-center" />
       <div className="hidden sm:absolute sm:-top-10 sm:right-1/2 sm:-z-10 sm:mr-10 sm:block sm:transform-gpu sm:blur-3xl" aria-hidden="true">
         <div className="aspect-[1097/845] w-[68.5625rem] bg-gradient-to-tr from-[#ff4694] to-[#776fff] opacity-20" style={{ clipPath: "polygon(74.1% 44.1%, 100% 61.6%, 97.5% 26.9%, 85.5% 0.1%, 80.7% 2%, 72.5% 32.5%, 60.2% 62.4%, 52.4% 68.1%, 47.5% 58.3%, 45.2% 34.5%, 27.5% 76.7%, 0.1% 64.9%, 17.9% 100%, 27.6% 76.8%, 76.1% 97.7%, 74.1% 44.1%)" }}></div>
       </div>
