+++ client/src/data/women-in-cs.js
@@ -3,7 +3,7 @@ export const womenInCS = [
   { firstName: "Ada", lastName: "Lovelace (Countess of Lovelace)", summary: "A British mathematician and writer, she is often regarded as the first computer programmer." },
   { firstName: "Grace", lastName: "Hopper", summary: "A computer scientist and naval officer who developed the first compiler for computer programming language." },
   { firstName: "Ida", lastName: "Rhodes", summary: "A mathematician and computer pioneer who worked on the development of the first electronic computer, ENIAC." },
-  { firstName: "Adele", lastName: "Goldberg", summary: "An American computer scientist who made significant contributions to the design of the Smalltalk programming language and the development of object-oriented programming." },
+  { firstName: "Adele", lastName: "Goldberg", summary: "An American computer scientist who made significant contributions to the design of the Smalltalk programming language and the infrastructure and design for overlapping windows on display screens, or 'Graphical User Interface' (GUI) as well as the development of object-oriented programming." },
   { firstName: "Lynn", lastName: "Conway", summary: "An American computer scientist and electrical engineer who made groundbreaking contributions to VLSI (Very Large Scale Integration) design and teaching." },
   { firstName: "Hedy", lastName: "Lamarr", summary: "Perhaps best known as a glamorous Hollywood actress, Hedy Lamarr was also an inventor. She developed a radio guidance system for Allied torpedoes during World War II, a technology that became a basis for modern Wi-Fi, Bluetooth, and GPS communication." },
   { firstName: "Radia", lastName: "Perlman", summary: "An American software designer and network engineer, often referred to as the \"Mother of the Internet.\" She designed the Spanning-Tree Protocol (STP), which is fundamental to the operation of network bridges." },
@@ -15,5 +15,10 @@ export const womenInCS = [
   { firstName: "Fei-Fei", lastName: "Li", summary: "Co-creator of ImageNet and Professor of Computer Science at Stanford University." },
   { firstName: "Shafi", lastName: "Goldwasser", summary: "A computer scientist and mathematician known for her work in cryptography and complexity theory. She is a recipient of the Turing Award, often referred to as the \"Nobel Prize of Computing.\"" },
   { firstName: "Karen", lastName: "Spärck Jones", summary: "A British computer scientist who made significant contributions to natural language processing and information retrieval. She is best known for her work on inverse document frequency (IDF) in information retrieval, which is still widely used today." },
+  { firstName: "Mary Allen", lastName: "Wilkes", summary: "An American computer programmer and lawyer. She was one of the first programmers of the LINC computer and the first person to use a computer at home. Wilkes is known for her contributions to the development of interactive computer systems and was the first to use a computer word processor for legal work." },
+  { firstName: "Reshma", lastName: "Saujani", summary: "An American lawyer, educator, and politician. She is the founder of Girls Who Code, a non-profit organization working to close the gender gap in technology and provide computer science education to young women. Saujani has played a pivotal role in inspiring and empowering a new generation of women in technology." },
+  { firstName: "Joan", lastName: "Ball", summary: "An English entrepreneur who founded Comet Computer, the first computer dating service, in 1964, which translated survey answers to punch cards that were ran through a time-share computer." },
+  { firstName: "Carol", lastName: "Shaw", summary: "A groundbreaking video game designer and programmer, known for her work on games like 'River Raid' and '3-D Tic-Tac-Toe'. She was the first female video game developer at Atari. Her unpublished 1978 Polo is the first documented game designed and programmed by a woman." },
+  { firstName: "Elizabeth", lastName: "Feinler", summary: "An information scientist who directed the Network Information Systems Center, which managed early internet infrastructure, including the ARPANET and NIC, where she helped introduce domain naming protocol." },
   { firstName: "Marie", lastName: "Curie", summary: "A Polish-born physicist and chemist, Marie Curie is famous for her pioneering research on radioactivity. She was one of the first women to win a Nobel Prize, and her work laid the foundation for future developments in X-ray technology and cancer treatment." }
 ]
